change rating to a float

+++ app/src/main/java/com/example/vitabu/UserAbstract.java
@@ -35,8 +35,8 @@ public abstract class UserAbstract extends Object{
     protected String userid;
     protected Location location;
     protected String userName;
-    protected int borrowerRating;
-    protected int ownerRating;
+    protected float borrowerRating;
+    protected float ownerRating;
     protected int booksOwned;
     protected int booksBorrowed;
     protected Date joinDate;
@@ -48,10 +48,10 @@ public abstract class UserAbstract extends Object{
     public abstract void setLocation(Location l);
     public abstract String getUserName();
     public abstract void setUserName(String s);
-    public abstract int getBorrowerRating();
-    public abstract void setBorrowerRating(int i);
-    public abstract int getOwnerRating();
-    public abstract void setOwnerRating(int i);
+    public abstract float getBorrowerRating();
+    public abstract void setBorrowerRating(float i);
+    public abstract float getOwnerRating();
+    public abstract void setOwnerRating(float i);
     public abstract int getBooksOwned();
     public abstract void setBooksOwned(int i);
     public abstract int getBooksBorrowed();
